Finalize central error handling and implement logger.js & validation.js middlewares

+++ controllers/clothingitems.js
@@ -1,50 +1,70 @@
 const clothingItem = require("../models/clothingitems");
+const {
+  BadRequestError,
+  NotFoundError,
+  ForbiddenError,
+} = require("../middlewares/error-handler");
 
-const getItems = (req, res) => {
+const getItems = (req, res, next) => {
   clothingItem
     .find({})
     .then((items) => res.status(200).send(items))
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send({ message: "Server error" });
-    });
+    .catch(next);
 };
 
-const postNewItem = (req, res) => {
+const postNewItem = (req, res, next) => {
   const { name, weather, imageUrl } = req.body;
   const owner = req.user._id;
+
+  if (!name || !weather || !imageUrl) {
+    throw new BadRequestError("Name, weather, and imageUrl are required");
+  }
 
   clothingItem
     .create({ name, weather, imageUrl, owner })
     .then((item) => res.status(201).send(item))
     .catch((err) => {
-      console.error(err);
-      res.status(400).send({ message: "Invalid data" });
+      if (err.name === "ValidationError") {
+        next(new BadRequestError(err.message));
+      } else {
+        next(err);
+      }
     });
 };
 
-const deleteItem = (req, res) => {
+const deleteItem = (req, res, next) => {
   const { itemId } = req.params;
+
+  if (!itemId) {
+    throw new BadRequestError("Item ID is required");
+  }
 
   clothingItem
     .findById(itemId)
     .then((item) => {
       if (!item) {
-        return res.status(404).send({ message: "Item not found" });
+        throw new NotFoundError("Item not found");
       }
       if (item.owner.toString() !== req.user._id) {
-        return res
-          .status(403)
-          .send({ message: "This item does not belong to you" });
+        throw new ForbiddenError(
+          "You don't have permission to delete this item"
+        );
       }
 
       return clothingItem
         .findByIdAndDelete(itemId)
-        .then(() => res.status(200).send({ message: "Item deleted " }));
+        .then(() =>
+          res.status(200).send({ message: "Item deleted successfully" })
+        );
     })
     .catch((err) => {
-      console.error(err);
-      res.status(500).send({ message: "Internal server error" });
+      if (err.name === "CastError") {
+        next(new BadRequestError("Invalid item ID format"));
+      } else if (err.statusCode) {
+        next(err);
+      } else {
+        next(err);
+      }
     });
 };
